+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Run the tests on a custom Elasticsearch port, fix #56
+* Improve Exception message when Factory is used badly
 
 ## [1.5.1] - 2022-01-18